Refactor Wildcard namespace and remove MG.Http.Urls
Moved Wildcard struct from MG.Sonarr.Next.Shell.Components to MG.Sonarr.Next.Strings. Updated using directives in multiple files to reflect the new namespace. Removed MG.Http.Urls package reference from MG.Sonarr.Next.csproj. Wildcard implementation remains unchanged.
