+++ src/MG.Sonarr.Next.Shell.Bases/Extensions/ArrayExtensions.cs
@@ -1,4 +1,5 @@
 ﻿using MG.Sonarr.Next.Shell.Components;
+using MG.Sonarr.Next.Strings;
 
 namespace MG.Sonarr.Next.Shell.Extensions
 {
