+++ src/MG.Sonarr.Next.Shell.Bases/Extensions/PSCmdletExtensions.cs
@@ -2,7 +2,7 @@
 using MG.Sonarr.Next.Extensions;
 using MG.Sonarr.Next.Reflection;
 using MG.Sonarr.Next.Shell.Cmdlets;
-using MG.Sonarr.Next.Shell.Components;
+using MG.Sonarr.Next.Strings;
 using System.Runtime.CompilerServices;
 
 namespace MG.Sonarr.Next.Shell.Extensions
