+++ src/MG.Sonarr.Next.Shell/Cmdlets/Calendar/GetSonarrCalendarCmdlet.cs
@@ -1,5 +1,4 @@
-﻿using MG.Http.Urls.Queries;
-using MG.Sonarr.Next.Metadata;
+﻿using MG.Sonarr.Next.Metadata;
 using MG.Sonarr.Next.Models.Calendar;
 using MG.Sonarr.Next.Shell.Extensions;
 using MG.Sonarr.Next.Extensions;
