+++ src/MG.Sonarr.Next.Shell/Cmdlets/Commands/TestSonarrRenameCmdlet.cs
@@ -1,6 +1,6 @@
-﻿using MG.Http.Urls.Queries;
-using MG.Sonarr.Next.Attributes;
+﻿using MG.Sonarr.Next.Attributes;
 using MG.Sonarr.Next.Extensions;
+using MG.Sonarr.Next.Extensions.Strings;
 using MG.Sonarr.Next.Metadata;
 using MG.Sonarr.Next.Models.Renames;
 using MG.Sonarr.Next.Shell.Attributes;
