+++ src/MG.Sonarr.Next.Shell/Cmdlets/Episodes/GetSonarrEpisodeFileCmdlet.cs
@@ -1,5 +1,4 @@
-﻿using MG.Http.Urls.Queries;
-using MG.Sonarr.Next.Attributes;
+﻿using MG.Sonarr.Next.Attributes;
 using MG.Sonarr.Next.Collections.Pools;
 using MG.Sonarr.Next.Metadata;
 using MG.Sonarr.Next.Models.Episodes;
