+++ src/MG.Sonarr.Next.Shell/Cmdlets/ManualImports/GetSonarrManualImportCmdlet.cs
@@ -1,4 +1,3 @@
-using MG.Http.Urls.Queries;
 using MG.Sonarr.Next.Metadata;
 using MG.Sonarr.Next.Models.ManualImports;
 using MG.Sonarr.Next.Models.Series;
