+++ src/MG.Sonarr.Next.Shell/Cmdlets/Releases/SearchSonarrReleaseCmdlet.cs
@@ -1,5 +1,4 @@
-﻿using MG.Http.Urls.Queries;
-using MG.Sonarr.Next.Attributes;
+﻿using MG.Sonarr.Next.Attributes;
 using MG.Sonarr.Next.Metadata;
 using MG.Sonarr.Next.Models.Releases;
 using MG.Sonarr.Next.Shell.Attributes;
