+++ src/MG.Sonarr.Next.Shell/Cmdlets/Series/SearchSonarrSeriesCmdlet.cs
@@ -1,7 +1,7 @@
 ﻿using MG.Sonarr.Next.Metadata;
 using MG.Sonarr.Next.Models.Series;
-using MG.Sonarr.Next.Shell.Components;
 using MG.Sonarr.Next.Shell.Extensions;
+using MG.Sonarr.Next.Strings;
 using System.Net;
 
 namespace MG.Sonarr.Next.Shell.Cmdlets.Series
