+++ src/MG.Sonarr.Next.Shell/Cmdlets/Systems/Logs/GetSonarrLogCmdlet.cs
@@ -1,5 +1,4 @@
-﻿using MG.Http.Urls.Queries;
-using MG.Sonarr.Next.Metadata;
+﻿using MG.Sonarr.Next.Metadata;
 using MG.Sonarr.Next.Models;
 using MG.Sonarr.Next.Models.System;
 using MG.Sonarr.Next.Services.Http.Queries;
