+++ src/MG.Sonarr.Next.Shell/Cmdlets/WantedMissing/GetSonarrWantedMissingCmdlet.cs
@@ -1,5 +1,4 @@
-﻿using MG.Http.Urls.Queries;
-using MG.Sonarr.Next.Models;
+﻿using MG.Sonarr.Next.Models;
 using MG.Sonarr.Next.Models.Episodes;
 using MG.Sonarr.Next.Shell.Extensions;
