+++ src/MG.Sonarr.Next/Extensions/Strings/SpanExtensions.cs
@@ -1,5 +1,4 @@
-﻿using MG.Sonarr.Next.Attributes;
-using MG.Sonarr.Next.Extensions.Strings;
+﻿using MG.Sonarr.Next.Extensions.Strings;
 
 namespace MG.Sonarr.Next.Extensions.Strings
 {
