+++ src/MG.Sonarr.Next/Metadata/MetadataTag.cs
@@ -1,5 +1,5 @@
-﻿using MG.Http.Urls.Queries;
-using MG.Sonarr.Next.Extensions;
+﻿using MG.Sonarr.Next.Extensions;
+using MG.Sonarr.Next.Extensions.Strings;
 
 namespace MG.Sonarr.Next.Metadata
 {
