+++ src/MG.Sonarr.Next/Services/Http/Clients/SignalRClient.cs
@@ -1,5 +1,4 @@
-﻿using MG.Http.Urls.Queries;
-using MG.Sonarr.Next.Services.Auth;
+﻿using MG.Sonarr.Next.Services.Auth;
 using MG.Sonarr.Next.Collections;
 using MG.Sonarr.Next.Collections.Pools;
 using MG.Sonarr.Next.Extensions;
@@ -10,6 +9,7 @@ using MG.Sonarr.Next.Services.Http.Requests;
 using Microsoft.Extensions.DependencyInjection;
 using System.Management.Automation;
 using System.Net.Http.Json;
+using MG.Sonarr.Next.Extensions.Strings;
 
 namespace MG.Sonarr.Next.Services.Http.Clients
 {
